Retire the last browser->PLoT /v2/run caller: untangle useV2Run (ROADMAP 2.1229)
Deletes the direct browser->PLoT analysis seam. The proxy allowlist derives
8 -> 7, by deleting the CALLER first and letting the entry fall out.

WHAT WENT
  - src/canvas/hooks/useV2Run.ts (whole hook)
  - adapter.ts "HTTP Client Integration": V2AdapterConfig, runV2 (the single
    /v2/run fetch), executeV2RunWithAnalysisReady, and their index exports
  - isAnalysisReadyStale (sole live caller was the hook's request builder)
  - { status: 'v2' } from CanonicalRunOutcome + its three consumer arms
  - OutputsDock's Cancel button: gated on the hook's own in-flight flag
    because cancelRun could only abort the direct request. Dead by
    construction now, and REMOVED rather than re-gated on the store status —
    a Cancel that cannot cancel a CEE turn is a false control.

WHAT STAYED
  Everything above the deleted section in adapter.ts — buildV2Request,
  validateAllEdges, transformEdgeToV2*, responseMapper, types — has live
  consumers that never touched the transport. Deleting a transport is not a
  reason to delete the shapes.

EXTRACTED
  src/canvas/hooks/goalThresholdResolvers.ts — the pure goal-threshold and
  goal-node resolvers, moved VERBATIM. Their consumers (OutputsDock,
  DefineSuccessModal, the inspector contract test) were never the direct-run
  path, so deleting the hook must not take them with it.

PREMISE CORRECTIONS (the row's blocker is largely refuted at the tip)
  - resultsStart is NOT the only writer of results.startedAt: resultsAnalysing,
    the canonical opener, already stamps it (store.ts:3633). Same for runEpoch,
    the lens reset and analysisStateReady.
  - results.seed CANNOT move to the canonical path. mapV5AnalysisToReport.ts
    :843-877 — "The V5 contract carries NO seed field ... seed is a provenance
    lie (T2 receipts-honesty)"; receipts fail closed to null, never 0. So the
    briefed "the seed round-trips" test is not satisfiable honestly.
  - A UI run-history writer on the canonical path is FORBIDDEN by ROADMAP
    2.1231 ("delta computed producer-side, never from UI run history") and
    duplicates 2.99 ("Fix is producer-boundary ... not UI"). Not built here.
  - Run history, the Supabase analysis write and setAnalysisSummary were
    ALREADY dark on the deployed canonical path, so nothing user-reachable is
    lost: the direct seam only ran when the canonical flag was OFF, and
    deployed posture is canonical ON (netlify.toml:154).

UNTOUCHED
  The ROADMAP 2.1127 error-mapping semantics went with the hook; they were not
  relocated, re-shaped or re-tuned. That redesign remains 2.1127's.
