+++ scripts/ci/typecheck-baseline.txt
@@ -10,7 +10,7 @@
 # this file and phase 2.
 #
 # Format: <error-count><TAB><path>, sorted by path.
-# count=2348
+# count=2325
 1	e2e/_helpers.ts
 1	e2e/a11y.spec.ts
 9	e2e/canvas-inspector.spec.ts
@@ -88,8 +88,7 @@
 4	src/adapters/plot/v2/__tests__/responseMapper.constraintGate.spec.ts
 14	src/adapters/plot/v2/__tests__/responseMapper.displayHonesty.spec.ts
 12	src/adapters/plot/v2/__tests__/responseMapper.spec.ts
-1	src/adapters/plot/v2/__tests__/runV2.plotBearer.spec.ts
-9	src/adapters/plot/v2/adapter.ts
+3	src/adapters/plot/v2/adapter.ts
 14	src/adapters/plot/v2/responseMapper.ts
 12	src/canvas/ReactFlowGraph.tsx
 3	src/canvas/__tests__/__helpers__/renderCanvas.tsx
@@ -165,7 +164,6 @@
 1	src/canvas/components/__tests__/ModelCardLite.spec.tsx
 2	src/canvas/components/__tests__/ModelTabBody.spec.tsx
 1	src/canvas/components/__tests__/ObjectiveBanner.spec.tsx
-1	src/canvas/components/__tests__/OutputsDock.dom.spec.tsx
 1	src/canvas/components/__tests__/RiskProfileBadge.spec.tsx
 7	src/canvas/components/__tests__/RiskProfileSelector.spec.tsx
 1	src/canvas/components/__tests__/RiskTolerancePanel.spec.tsx
@@ -242,11 +240,8 @@
 1	src/canvas/health/issueExplainers.ts
 2	src/canvas/highlight/HighlightLayer.tsx
 1	src/canvas/hooks/__tests__/goalThreshold.representation.spec.ts
-2	src/canvas/hooks/__tests__/isAnalysisReadyStale.spec.ts
 2	src/canvas/hooks/__tests__/useGuidancePulseHighlight.spec.ts
 13	src/canvas/hooks/__tests__/usePathHighlight.spec.ts
-2	src/canvas/hooks/__tests__/useV2Run.concurrency.spec.ts
-1	src/canvas/hooks/__tests__/useV2Run.interventionValidation.spec.ts
 1	src/canvas/hooks/useAddBaseline.ts
 5	src/canvas/hooks/useAnalysisMetadata.ts
 1	src/canvas/hooks/useAutosave.ts
@@ -265,7 +260,6 @@
 9	src/canvas/hooks/useScienceIcons.ts
 20	src/canvas/hooks/useUnifiedActions.ts
 6	src/canvas/hooks/useUtilityWeights.ts
-10	src/canvas/hooks/useV2Run.ts
 2	src/canvas/nodes/ActionNode.tsx
 1	src/canvas/nodes/BaseNode.tsx
 6	src/canvas/nodes/DecisionNode.tsx
