+++ src/canvas/components/__tests__/OutputsDock.announcerPlacement.spec.tsx
@@ -26,12 +26,10 @@ import { ConversationProvider } from '../../conversation/ConversationContext'
 const {
   mockIsV5CanonicalAnalysisEnabled,
   mockIsV5Eligible,
-  mockUseV2Run,
   mockShowToast,
 } = vi.hoisted(() => ({
   mockIsV5CanonicalAnalysisEnabled: vi.fn(() => false),
   mockIsV5Eligible: vi.fn((_input?: { flag: string | undefined }) => ({ eligible: false, reason: 'flag_off' })),
-  mockUseV2Run: vi.fn(() => ({ runV2Analysis: vi.fn(), cancelRun: vi.fn() })),
   mockShowToast: vi.fn(),
 }))
 
@@ -60,11 +58,6 @@ vi.mock('../../../v5/eligibility', async (importOriginal) => {
       flags.isV5CanonicalAnalysisEnabled() &&
       mockIsV5Eligible({ flag: import.meta.env.VITE_ENABLE_V5_ORCHESTRATOR }).eligible,
   }
-})
-
-vi.mock('../../hooks/useV2Run', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../../hooks/useV2Run')>()
-  return { ...actual, useV2Run: () => mockUseV2Run() }
 })
 
 vi.mock('../../ToastContext', async (importOriginal) => {
@@ -159,7 +152,6 @@ describe('OutputsDock announcer placement (C3)', () => {
   beforeEach(() => {
     ensureMatchMedia()
     vi.clearAllMocks()
-    mockUseV2Run.mockReturnValue({ runV2Analysis: vi.fn(), cancelRun: vi.fn() })
     seedCompletedRun()
     useUIStore.setState({ activeRightPanel: null } as never)
   })
