+++ src/canvas/components/__tests__/OutputsDock.assistantOpenedAttribution.spec.tsx
@@ -57,7 +57,6 @@ import { ConversationProvider } from '../../conversation/ConversationContext'
 const {
   mockIsV5CanonicalAnalysisEnabled,
   mockIsV5Eligible,
-  mockUseV2Run,
   mockShowToast,
   mockIsJourneyTabEnabled,
   mockIsCompareTabEnabled,
@@ -65,7 +64,6 @@ const {
 } = vi.hoisted(() => ({
   mockIsV5CanonicalAnalysisEnabled: vi.fn(() => false),
   mockIsV5Eligible: vi.fn((_input?: { flag: string | undefined }) => ({ eligible: false, reason: 'flag_off' })),
-  mockUseV2Run: vi.fn(() => ({ runV2Analysis: vi.fn(), cancelRun: vi.fn() })),
   mockShowToast: vi.fn(),
   // ⚠ TRAP 3b — THESE DEFAULTS ARE THE DEPLOYED POSTURE, NOT A CONVENIENCE.
   // Derived 13 Aug 2026 from `netlify.toml` `[context.staging.environment]`
@@ -108,11 +106,6 @@ vi.mock('../../../v5/eligibility', async (importOriginal) => {
       flags.isV5CanonicalAnalysisEnabled() &&
       mockIsV5Eligible({ flag: import.meta.env.VITE_ENABLE_V5_ORCHESTRATOR }).eligible,
   }
-})
-
-vi.mock('../../hooks/useV2Run', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../../hooks/useV2Run')>()
-  return { ...actual, useV2Run: () => mockUseV2Run() }
 })
 
 vi.mock('../../ToastContext', async (importOriginal) => {
@@ -303,7 +296,6 @@ describe('ROADMAP 2.1132 — the assistant attributes the panel gestures it actu
   beforeEach(() => {
     ensureMatchMedia()
     vi.clearAllMocks()
-    mockUseV2Run.mockReturnValue({ runV2Analysis: vi.fn(), cancelRun: vi.fn() })
     mockIsJourneyTabEnabled.mockReturnValue(false)
     mockIsCompareTabEnabled.mockReturnValue(true)
     mockIsAiPanelV2Enabled.mockReturnValue(true)
