+++ src/canvas/components/__tests__/OutputsDock.overlayOcclusion.spec.tsx
@@ -51,12 +51,10 @@ import { ConversationProvider } from '../../conversation/ConversationContext'
 const {
   mockIsV5CanonicalAnalysisEnabled,
   mockIsV5Eligible,
-  mockUseV2Run,
   mockShowToast,
 } = vi.hoisted(() => ({
   mockIsV5CanonicalAnalysisEnabled: vi.fn(() => false),
   mockIsV5Eligible: vi.fn((_input?: { flag: string | undefined }) => ({ eligible: false, reason: 'flag_off' })),
-  mockUseV2Run: vi.fn(() => ({ runV2Analysis: vi.fn(), cancelRun: vi.fn() })),
   mockShowToast: vi.fn(),
 }))
 
@@ -85,11 +83,6 @@ vi.mock('../../../v5/eligibility', async (importOriginal) => {
       flags.isV5CanonicalAnalysisEnabled() &&
       mockIsV5Eligible({ flag: import.meta.env.VITE_ENABLE_V5_ORCHESTRATOR }).eligible,
   }
-})
-
-vi.mock('../../hooks/useV2Run', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../../hooks/useV2Run')>()
-  return { ...actual, useV2Run: () => mockUseV2Run() }
 })
 
 vi.mock('../../ToastContext', async (importOriginal) => {
@@ -223,7 +216,6 @@ describe('ROADMAP 2.639 — a panel directive is not executed behind an overlay'
   beforeEach(() => {
     ensureMatchMedia()
     vi.clearAllMocks()
-    mockUseV2Run.mockReturnValue({ runV2Analysis: vi.fn(), cancelRun: vi.fn() })
     seedCompletedRun()
     useUIStore.setState({
       activeRightPanel: null,
