+++ src/canvas/components/__tests__/OutputsDock.testability.spec.tsx
@@ -44,10 +44,6 @@ vi.mock('../../../flags', async (importOriginal) => {
   }
 })
 
-vi.mock('../../hooks/useV2Run', () => ({
-  useV2Run: () => ({ runV2Analysis: vi.fn(), cancelRun: vi.fn() }),
-}))
-
 // Stable conversation stub (shape mirrors the singleton spec's stub).
 vi.mock('../../conversation/useConversation', () => ({
   useConversation: () => ({
