--- src/canvas/hooks/useV2Run.ts
@@ -1,1268 +1,0 @@
-/**
- * V2 Run Hook (P0-UI Integration)
- *
- * Replaces useResultsRun for V2 adapter integration.
- * Calls /v2/run and maps response to store format.
- */
-
-import { useCallback, useRef, useState } from 'react'
-import { useCanvasStore } from '../store'
-import { useDraftStore } from '../stores/draftStore'
-import { useResultsStore } from '../stores/resultsStore'
-import { getEdgeKey } from '../domain/edgeUtils'
-import {
-  executeV2RunWithAnalysisReady,
-  isBlockedResponse,
-  isFailedAnalysis,
-  isSuccessfulAnalysis,
-  validateV2RunResponseFull,
-  sanitizeV2RunResponse,
-  reconcileOptionsWithCanvasNodesDetailed,
-  flattenInterventions,
-  clearStrengthCorrections,
-  type V2AdapterConfig,
-  type V2RunError,
-  type V2RunResponse,
-} from '../../adapters/plot/v2'
-import {
-  mapV2ResponseToReportV1,
-  createErrorReport,
-  createEnrichmentFromV2Response,
-  detectComputedButEmpty,
-  synthesizeCeeReviewFromV2,
-  synthesizeCeeTraceFromV2,
-} from '../../adapters/plot/v2/responseMapper'
-// ⚠ RUN-SPINE TELEMETRY DELIBERATELY DOES NOT LIVE HERE — ROADMAP 1.68.
-//
-// This hook used to call `trackRunCompleted` once and `trackRunFailed` four
-// times. `OutputsDock.tsx` ALSO emits both, from a store-status-transition
-// effect — and OutputsDock consumes this very hook (`OutputsDock.tsx:696`), so
-// both always fired: TWO `run_completed` per run, with two DISJOINT payload
-// shapes (this hook sent `{duration_ms, option_count, has_drivers, request_id}`;
-// OutputsDock sends the declared `{confidence_level, drivers_informative,
-// trace_id, duration_ms}`). One event name with two shapes is the precise
-// defect the measurement design rejected `metrics.ts` for, and it was live.
-//
-// OutputsDock won as the single canonical emitter, on COVERAGE — not on
-// seniority. It observes the results STORE, so it fires for every run path:
-// this hook, `useResultsRun.ts`, `applyV5State.ts:1064` and
-// `useConversation.ts:3144/:3251`. The last two are the CEE-driven analysis
-// path, which never goes through this hook — so deleting OutputsDock's emitters
-// instead would have silently dropped the platform's primary analysis path.
-// `<OutputsDock />` is mounted unconditionally at `ReactFlowGraph.tsx:2243` and
-// collapses via CSS, so its effect is always live.
-//
-// A per-run dedup latch inside the seam was considered and REJECTED: there is
-// no reliable per-run identity in the store to key it on. `results.runId` is
-// written only by `resultsConnecting`, whose sole caller is
-// `useResultsRun.ts:85`; `results.startedAt` is written only by `resultsStart`,
-// called from this hook and `useResultsRun` but NOT from the CEE path. Keying a
-// latch on either would work in tests and fail silently on the path that
-// matters most.
-//
-// Pinned by `src/lib/__tests__/runSpineSingleEmission.spec.ts`.
-import { trackEmptyComputedResults } from '../../lib/resultsInstrumentation'
-import { generateGraphHash } from '../store/runHistory'
-import { trackTypedError } from '../../lib/telemetry'
-import { ApiError, NetworkError, ProcessingError, isApiError } from '../../lib/api-errors'
-import { generateRequestId } from '../../types/requestId'
-import type { CEEAnalysisReady } from '../../adapters/cee/types'
-import type { M1Review, M1Coaching } from '../../types/cee'
-import { useGateStore, updateRobustnessGate, updateRobustnessGateFromV2 } from '../../lib/gate-state'
-import { buildRawErrorData, hashStackTrace } from '../../utils/payloadRedaction'
-import { extractM1ReviewFromV2, extractM1CoachingFromV2 } from '../../hooks/hydrateAnalysis'
-import { assembleAnalysisInputsSummary } from '../analysis/assembleAnalysisInputsSummary'
-import {
-  useSuccessMeasureStore,
-  selectSuccessMeasure,
-} from '../../components/results/modals/successMeasureStore'
-import { resolveScenarioKey } from '../../components/results/modals/scenarioKey'
-
-/**
- * P0 Fix: Derive a stable numeric seed from a string.
- * Uses simple FNV-1a hash to convert string to number.
- */
-function deriveNumericSeedFromString(input: string): number {
-  let hash = 2166136261 // FNV offset basis
-  for (let i = 0; i < input.length; i++) {
-    hash ^= input.charCodeAt(i)
-    hash = Math.imul(hash, 16777619) // FNV prime
-  }
-  // Convert to positive number in reasonable seed range (0 - 999999)
-  return Math.abs(hash) % 1000000
-}
-
-/**
- * Re-exported from './resolveSeedUsed'. It was moved to a pure leaf module so a
- * second caller could import it without being blanked by the hand-listed
- * `vi.mock` factories that replace THIS module wholesale in a dozen specs (and
- * without dragging the whole hook's dependency graph into their chunk). One
- * implementation, unchanged behaviour.
- */
-import { resolveSeedUsed } from './resolveSeedUsed'
-export { resolveSeedUsed }
-
-
-/**
- * UI-SEM-058: raw → normalised goal-threshold conversion for the PLoT request.
- *
- * store.goalThreshold holds USER UNITS by default (raw), but since Lane 5 its
- * representation is carried explicitly by store.goalThresholdRepresentation
- * ('raw' | 'normalised' | null) — a CEE bare-sync can store an already-0-1
- * value tagged 'normalised', in which case resolveChipGoalThreshold
- * short-circuits and never divides by a cap. PLoT's `goal_threshold` contract
- * is normalised 0-1 (adapter.ts request builder). Convert raw/cap when the CEE
- * cap exists.
- * Without a cap, raw ≡ normalised only when the value already lies in [0,1].
- * Anything that cannot be proven normalised is OMITTED (returns undefined)
- * so the request builder's own `analysisReady.goal_threshold` (already
- * normalised) stands — a missing threshold degrades to "no probability_of_goal",
- * never to a corrupt analysis. Format conversion only (same class as UI-SEM-001).
- */
-export function normaliseGoalThresholdForRequest(
-  raw: number | null | undefined,
-  cap: unknown,
-): number | undefined {
-  if (raw == null || !Number.isFinite(raw)) return undefined
-  const capNumber = typeof cap === 'number' && Number.isFinite(cap) && cap > 0 ? cap : null
-  let normalised = capNumber != null ? raw / capNumber : raw
-  // Floating-point guard: a threshold set exactly at the cap must normalise
-  // to 1.0 even when the division lands one ULP above it — a legitimate
-  // 100% target must never be dropped over the last bit of a float.
-  if (normalised > 1 && normalised < 1 + 1e-9) normalised = 1
-  if (normalised < 0 || normalised > 1) {
-    console.warn(
-      '[useV2Run] goal threshold override omitted — cannot prove normalised form',
-      { raw, cap: capNumber, normalised },
-    )
-    return undefined
-  }
-  return normalised
-}
-
-/**
- * Lane 5 (Codex P1) — the ONE goal-node resolution used for request
- * construction, cap lookup and threshold commitment across V2 and V5, so the
- * legs can never resolve different nodes. Precedence: the CEE-certified
- * `analysis_ready.goal_node_id`, then the store's `outcomeNodeId`, then the
- * first goal node — but each candidate is validated to EXIST in the current
- * graph, so a stale `outcomeNodeId` (retained across a scenario replacement)
- * is skipped rather than silently resolved to a reseeded-away id.
- *
- * Before this, the V2 cap used `outcomeNodeId` while the adapter + V5 chip
- * preferred `analysis_ready.goal_node_id`: with divergent ids and caps, V2
- * could normalise against one node's cap while the request named the other.
- */
-export function resolveActiveGoalNodeId(state: {
-  ceeAnalysisReady?: { goal_node_id?: string | null } | null
-  outcomeNodeId?: string | null
-  nodes: ReadonlyArray<{ id: string; type?: string; data?: unknown }>
-}): string | null {
-  const exists = (id: string | null | undefined): id is string =>
-    !!id && state.nodes.some((n) => n.id === id)
-  if (exists(state.ceeAnalysisReady?.goal_node_id)) return state.ceeAnalysisReady!.goal_node_id!
-  if (exists(state.outcomeNodeId)) return state.outcomeNodeId!
-  const goalNode = state.nodes.find(
-    (n) => n.type === 'goal' || (n.data as { type?: string } | undefined)?.type === 'goal',
-  )
-  return goalNode?.id ?? null
-}
-
-/**
- * Resolve the goal-threshold scale cap for the request boundary using the
- * SAME fallback chain the display path uses (useResultsSectionData
- * goalThresholdCap: analysis_ready first, then the goal node's data — CEE
- * fields are spread onto the node by DraftChat). Without the node fallback
- * the two boundaries could disagree: the display would denormalise outcomes
- * against a node-held cap while the request boundary, blind to it, dropped
- * the user's threshold override.
- */
-export function resolveGoalThresholdCap(
-  analysisReady: { goal_threshold_cap?: unknown } | null,
-  nodes: ReadonlyArray<{ id: string; data?: unknown }>,
-  goalNodeId: string | null | undefined,
-): number | undefined {
-  const goalNode = goalNodeId ? nodes.find((n) => n.id === goalNodeId) : undefined
-  const data = goalNode?.data as
-    | { goal_threshold_cap?: unknown; threshold_cap?: unknown; scale_max?: unknown }
-    | undefined
-  for (const candidate of [
-    analysisReady?.goal_threshold_cap,
-    data?.goal_threshold_cap,
-    data?.threshold_cap,
-    data?.scale_max,
-  ]) {
-    if (typeof candidate === 'number' && Number.isFinite(candidate) && candidate > 0) {
-      return candidate
-    }
-  }
-  return undefined
-}
-
-/**
- * UI-SEM-081 — unit-derived goal-threshold cap. A "%" unit stated by the USER
- * (Define-success unit picker / saved success measure) is a definitional cap
- * of 100 — user input, not fabrication. Every other unit has no definitional
- * scale, so no cap is invented (fail-closed omission stands). Consulted only
- * as the LAST resort after the producer/node cap chain (UI-SEM-058).
- */
-export function capForUnit(unit: string | null | undefined): number | undefined {
-  return typeof unit === 'string' && unit.trim() === '%' ? 100 : undefined
-}
-
-/**
- * UI-SEM-081 provenance guard (adversarial-review blocker fold): the saved
- * measure's unit may cap ONLY a store threshold that provably CAME from that
- * measure (identical raw value). The store field can hold values from other
- * writers in other scales — CEE-sync writes capless NORMALISED values into it
- * (store.ts bare-sync), and the field survives scenario switches while the
- * measure is scenario-keyed — so pairing the measure's "%" with a foreign
- * store value ships a silently wrong wire number (0.6 → 0.006, a 100× target
- * shrink). No provenance → no unit cap (fail closed).
- */
-export function resolveMeasureUnitCap(
-  measure: { threshold: number; unit: string } | null | undefined,
-  rawThreshold: number | null | undefined,
-): number | undefined {
-  if (!measure || rawThreshold == null) return undefined
-  return measure.threshold === rawThreshold ? capForUnit(measure.unit) : undefined
-}
-
-/**
- * UI-SEM-058 (V5 leg) — normalise a goal_threshold for a CANONICAL/V5 chip
- * parameter. The store holds RAW user units; the V2 request builder converts
- * at its boundary, but the V5 path (buildPayload) forwards chip parameters
- * VERBATIM. So any chip that carries goal_threshold must carry the NORMALISED
- * 0-1 form, and OMIT it when normalisation cannot be proven — a raw value on
- * the V5 wire is silently ignored by PLoT while it retains a stale target
- * (Codex final-audit B3: entering 60% shipped goal_threshold:60, HTTP 200,
- * old 0.53 retained, all lenses unavailable). Returns undefined → the caller
- * must omit the parameter (fail closed), never send raw.
- *
- * ctx.unitCap (UI-SEM-081) is consulted only when the producer/node chain
- * yields no cap — live staging drafts carry neither `goal_threshold_cap` nor
- * `scale_max`, which silently swallowed every %-unit target (V-P0-1,
- * 2026-07-13 wire evidence).
- *
- * ctx.representation (Lane 5, Codex P0-1) makes the value's UNITS EXPLICIT
- * rather than inferred from cap availability. A value tagged 'normalised'
- * (the store's bare-sync writes CEE's already-0-1 goal_threshold) is passed
- * through iff ∈[0,1] and NEVER divided by any cap — inferring "raw because a
- * cap exists" divided a normalised 0.6 by the node's cap 100 → 0.006. A
- * 'raw' or absent tag keeps the cap-chain conversion (backward compatible).
- */
-export function resolveChipGoalThreshold(
-  rawThreshold: number | null | undefined,
-  ctx: {
-    analysisReady: { goal_threshold_cap?: unknown } | null
-    nodes: ReadonlyArray<{ id: string; data?: unknown }>
-    goalNodeId: string | null | undefined
-    unitCap?: number
-    representation?: 'raw' | 'normalised' | null
-  },
-): number | undefined {
-  // Lane 5: an explicitly-normalised value is already 0-1 — validate and
-  // pass through, never apply a cap (cap division is a raw→normalised op).
-  if (ctx.representation === 'normalised') {
-    return normaliseGoalThresholdForRequest(rawThreshold, undefined)
-  }
-  const chainCap = resolveGoalThresholdCap(ctx.analysisReady, ctx.nodes, ctx.goalNodeId)
-  const unitCap =
-    typeof ctx.unitCap === 'number' && Number.isFinite(ctx.unitCap) && ctx.unitCap > 0
-      ? ctx.unitCap
-      : undefined
-  return normaliseGoalThresholdForRequest(rawThreshold, chainCap ?? unitCap)
-}
-
-/**
- * Check if ceeAnalysisReady is stale (graph has changed since it was stored).
- *
- * Returns true if stale (should not use analysis_ready).
- * Checks:
- * - goal_node_id still exists in current nodes
- * - All intervention target node IDs still exist in current nodes
- * - No new option nodes have been added since analysisReady was captured
- * - Option labels have not changed
- */
-export function isAnalysisReadyStale(
-  analysisReady: CEEAnalysisReady,
-  currentNodeIds: Set<string>,
-  storedNodeIds: string[] | null,
-  currentNodes?: Array<{ id: string; data?: Record<string, unknown> }>
-): { isStale: boolean; reason?: string } {
-  // Check if stored node IDs are available for comparison
-  if (!storedNodeIds) {
-    // No stored IDs - can't validate, assume valid (backwards compatibility)
-    return { isStale: false }
-  }
-
-  // Check if goal node was deleted
-  if (!currentNodeIds.has(analysisReady.goal_node_id)) {
-    return { isStale: true, reason: 'goal node no longer exists' }
-  }
-
-  // Check if any nodes were deleted that affect the analysis
-  const deletedNodeIds = storedNodeIds.filter((id) => !currentNodeIds.has(id))
-
-  if (deletedNodeIds.length > 0) {
-    const affectedOptions: string[] = []
-
-    for (const option of analysisReady.options) {
-      const targetIds = Object.keys(option.interventions)
-      const deletedTargets = targetIds.filter((id) => deletedNodeIds.includes(id))
-      if (deletedTargets.length > 0) {
-        affectedOptions.push(option.label)
-      }
-    }
-
-    if (affectedOptions.length > 0) {
-      return { isStale: true, reason: `intervention targets deleted for: ${affectedOptions.join(', ')}` }
-    }
-  }
-
-  // Check if any intervention targets no longer exist
-  for (const option of analysisReady.options) {
-    for (const targetId of Object.keys(option.interventions)) {
-      if (!currentNodeIds.has(targetId)) {
-        return { isStale: true, reason: `intervention target "${targetId}" no longer exists` }
-      }
-    }
-  }
-
-  // Check if any option nodes in analysisReady were deleted from the canvas
-  for (const option of analysisReady.options) {
-    if (!currentNodeIds.has(option.id)) {
-      return { isStale: true, reason: `option node "${option.label}" (${option.id}) deleted` }
-    }
-  }
-
-  // Check for new option nodes added since analysisReady was captured
-  if (currentNodes) {
-    const arOptionIds = new Set(analysisReady.options.map((o) => o.id))
-    const currentOptionNodes = currentNodes.filter(
-      (n) => n.data?.kind === 'option' || n.data?.type === 'option'
-    )
-    const newOptions = currentOptionNodes.filter((n) => !arOptionIds.has(n.id))
-    if (newOptions.length > 0) {
-      const labels = newOptions.map((n) => (n.data?.label as string) || n.id).join(', ')
-      return { isStale: true, reason: `new option nodes added: ${labels}` }
-    }
-
-    // Check if option labels changed
-    for (const arOption of analysisReady.options) {
-      const currentNode = currentNodes.find((n) => n.id === arOption.id)
-      if (currentNode) {
-        const currentLabel = currentNode.data?.label as string | undefined
-        if (currentLabel && currentLabel !== arOption.label) {
-          return { isStale: true, reason: `option label changed: "${arOption.label}" → "${currentLabel}"` }
-        }
-      }
-    }
-  }
-
-  return { isStale: false }
-}
-
-/** Optional Supabase persistence callbacks injected by the calling component. */
-export interface V2RunPersistence {
-  setAnalysisRunning: () => Promise<void>
-  persistAnalysisSuccess: (
-    analysis: unknown,
-    graphHash: string,
-    /** T2b: null when the engine did not echo a usable seed — never a fabricated 0. */
-    seedUsed: number | null,
-    responseHash: string,
-    details?: Record<string, unknown>,
-  ) => Promise<void>
-  persistAnalysisFailure: (
-    errorPayload: { code: string; message: string },
-  ) => Promise<void>
-  /** Reset analysis_status to 'none' — called on user-initiated cancel. */
-  resetAnalysisStatus?: () => Promise<void>
-}
-
-interface UseV2RunReturn {
-  runV2Analysis: () => Promise<void>
-  cancelRun: () => void
-  isRunning: boolean
-  error: string | null
-}
-
-/**
- * Hook for running V2 analysis.
- *
- * Uses executeV2Run and maps response to store format.
- * Accepts optional Supabase persistence callbacks (C.1b Task 6).
- */
-export function useV2Run(persistence?: V2RunPersistence): UseV2RunReturn {
-  const [isRunning, setIsRunning] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const abortControllerRef = useRef<AbortController | null>(null)
-
-  // Audit F-76: Store selectors moved inside callback to prevent stale closures.
-  // Graph state is now snapshot at analysis trigger time, not render time.
-  // Store actions are stable references — safe to read from getState() inside callback.
-
-  const runV2Analysis = useCallback(async () => {
-    // Audit F-76: Snapshot graph state at trigger time (not render time)
-    const {
-      nodes,
-      edges,
-      outcomeNodeId,
-      goalThreshold,
-      goalThresholdRepresentation,
-      currentScenarioFraming: framing,
-      ceeAnalysisReady,
-      ceeAnalysisReadyNodeIds,
-      goalConstraints,
-      resultsStart,
-      resultsComplete,
-      resultsError,
-      resultsCancelled,
-      setRunMeta,
-      setCeeAnalysisReady,
-      setGoalConstraints,
-      captureErrorDetail,
-    } = useCanvasStore.getState()
-    const lastDraftDescription = useDraftStore.getState().lastDraftDescription
-    // Lane 3 review fold: the strength-corrections buffer is module-level
-    // and was never cleared, so the per-report snapshot (OutputsDock, keyed
-    // on [report]) accumulated duplicate rows across runs. Each V2 run
-    // records only its OWN corrections.
-    clearStrengthCorrections()
-
-    console.info('[constraint-trace] run-snapshot', {
-      source: 'useV2Run',
-      count: goalConstraints?.length ?? 0,
-      constraint_ids: goalConstraints?.map((c) => c.constraint_id) ?? [],
-      goal_threshold_present: goalThreshold != null,
-    })
-
-    // Validate goal is selected
-    if (!outcomeNodeId) {
-      setError('No goal node selected')
-      return
-    }
-
-    // Audit F-77: Abort any in-progress run before starting a new one.
-    // Prevents orphaned AbortControllers from double-click race conditions.
-    if (abortControllerRef.current && !abortControllerRef.current.signal.aborted) {
-      abortControllerRef.current.abort()
-    }
-
-    // Create a fresh AbortController for this run so the user can cancel
-    const controller = new AbortController()
-    abortControllerRef.current = controller
-
-    setIsRunning(true)
-    setError(null)
-
-    // P0 Fix: Seed precedence (no hardcoded "42" default)
-    // 1. Explicit seed from framing (user input or dev/test)
-    // 2. Derived from scenario_id if available (stable per scenario)
-    // 3. Derived from graph hash (stable per graph structure)
-    // 4. Let backend generate (omit seed from request)
-    let seed: number | undefined = framing?.seed
-
-    if (seed === undefined && framing?.scenario_id) {
-      // Derive stable seed from scenario_id hash
-      seed = deriveNumericSeedFromString(framing.scenario_id)
-    }
-
-    if (seed === undefined && nodes.length > 0) {
-      // Derive stable seed from graph structure hash
-      const graphKey = nodes.map(n => n.id).sort().join('|') + '::' + edges.map(e => getEdgeKey(e)).sort().join('|')
-      seed = deriveNumericSeedFromString(graphKey)
-    }
-
-    // If still undefined, let backend generate (don't use hardcoded 42)
-    // Use a timestamp-based seed as last resort for reproducibility tracking
-    if (seed === undefined) {
-      seed = Math.floor(Date.now() / 1000) % 1000000 // Use Unix timestamp modulo for reasonable range
-    }
-
-    const startTime = Date.now()
-
-    // Generate request ID for tracing
-    const requestId = generateRequestId()
-
-    if (import.meta.env.DEV) {
-      console.warn('[useV2Run] Request ID:', requestId)
-    }
-
-    // C.1b: Mark analysis as running in Supabase (fire-and-forget, non-blocking)
-    if (persistence) {
-      persistence.setAnalysisRunning().catch(() => {
-        // Non-critical — UI already shows running state
-      })
-    }
-
-    // Signal run start
-    resultsStart({ seed })
-    useResultsStore.getState().setAnalysisSummary(null)
-
-    // Reset run metadata
-    setRunMeta({
-      diagnostics: undefined,
-      correlationIdHeader: undefined,
-      degraded: undefined,
-      ceeReview: null,
-      ceeTrace: null,
-      ceeError: null,
-      ceeReviewV1: null,
-      ceeTraceV1: null,
-      ceeErrorV1: null,
-      // v2.0 diagnostic trace — clear stale data from prior runs
-      ceeDiagnosticTrace: null,
-    })
-
-    try {
-      // Get V2 adapter config
-      const config: V2AdapterConfig = {
-        baseUrl: import.meta.env?.VITE_PLOT_PROXY_BASE || '/bff/engine',
-        timeout: 120000,
-        signal: controller.signal,
-      }
-
-      // Check if ceeAnalysisReady is stale (graph changed since draft was applied)
-      let effectiveAnalysisReady = ceeAnalysisReady
-      let effectiveGoalConstraints = goalConstraints
-      const currentNodeIds = new Set(nodes.map((n) => n.id))
-
-      if (ceeAnalysisReady) {
-        const staleCheck = isAnalysisReadyStale(
-          ceeAnalysisReady,
-          currentNodeIds,
-          ceeAnalysisReadyNodeIds,
-          nodes
-        )
-
-        if (staleCheck.isStale) {
-          if (import.meta.env.DEV) {
-            console.warn('[useV2Run] Stale analysis_ready detected, falling back to node extraction', {
-              reason: staleCheck.reason,
-            })
-          }
-          // Clear stale analysis_ready and goal_constraints from store.
-          // System run-prep cleanup, not a user edit → do not self-dirty freshness.
-          setCeeAnalysisReady(null)
-          setGoalConstraints(null, { fromProducerSync: true })
-          effectiveAnalysisReady = null
-          effectiveGoalConstraints = null
-        }
-      }
-
-      // ── Pre-run intervention validation gate ──────────────────────────
-      // Validate that every option has at least one intervention before
-      // sending to PLoT. Empty interventions produce meaningless results.
-      //
-      // CRITICAL: this gate runs AFTER reconciliation so it sees the same
-      // option set the adapter will actually send. Reading from raw
-      // analysisReady.options here would block legitimate AI add_option
-      // flows whose interventions live on node.data.interventions.
-      // See reconcileOptionsWithCanvasNodes in src/adapters/plot/v2/adapter.ts.
-      {
-        // silent: true — the request-build path will call reconcile again and emit
-        // the canonical fallback console.warn. Avoid duplicate logs per run.
-        // phase: 'pre_run_check' — structured backfill telemetry still emits
-        // (with the phase tag so aggregators can dedupe across the two calls
-        // in the success path). Without the phase tag, blocked runs would
-        // hide all fallback usage because they return early below before the
-        // second reconcile call ever happens. See
-        // adapters/plot/v2/adapter.ts:ReconcileOptionsHookOptions.
-        // DETAILED form. The plain `reconcileOptionsWithCanvasNodes` returns only
-        // the options, and its canvas-backfill branches have ALREADY discarded any
-        // unusable entry by then — so a gate reading the options alone is
-        // structurally unable to see a partial loss on those branches. The detailed
-        // form reports what the walk removed. See ReconcileOptionsResult.
-        const { options: optionsToValidate, unusableByOptionId } =
-          reconcileOptionsWithCanvasNodesDetailed(
-            effectiveAnalysisReady,
-            nodes as any,
-            currentNodeIds,
-            { silent: true, phase: 'pre_run_check', scenarioId: framing?.scenario_id ?? null },
-          )
-
-        // Node labels for naming a target the way the user sees it on the canvas,
-        // not by node id — the convention PR #499 set for the same disposal.
-        const labelByNodeId = new Map<string, string>()
-        for (const n of nodes as any[]) {
-          const l = (n?.data as Record<string, unknown> | undefined)?.label
-          if (typeof l === 'string' && l.trim().length > 0) labelByNodeId.set(n.id, l)
-        }
-        const targetName = (id: string) => labelByNodeId.get(id) ?? id
-
-        // Identify offending options directly by id (not label) so duplicate
-        // labels do not collapse two distinct options into one entry.
-        // validateOptionsHaveInterventions returns labels only, which is fine
-        // for the human message but loses identity — we re-walk the reconciled
-        // list with the same usability rule (flattenInterventions, the single
-        // canonical "is this a usable map?" check) and key by opt.id.
-        //
-        // TWO failure modes, one affordance. Both end as MISSING_INTERVENTIONS +
-        // affectedOptions, which OutputsDock renders as the amber "Options need
-        // their effects mapped" banner with a focus button per option:
-        //
-        //   EMPTY   — no usable intervention at all. Pre-existing behaviour,
-        //             wording unchanged.
-        //   PARTIAL — some usable, at least one authored-but-unusable. NEW. This
-        //             used to run: the option was quietly shrunk to its usable
-        //             entries and the user got a confident answer about a graph
-        //             they never authored. Blocking is the doctrine — what gets
-        //             analysed is never silently altered — and blocking BEFORE the
-        //             request is built is what turns the adapter's typed throw
-        //             into an unreachable invariant rather than a crash.
-        const emptyOptions: Array<{ id: string; label: string }> = []
-        const partialOptions: Array<{ id: string; label: string; targets: string[] }> = []
-        for (const opt of optionsToValidate) {
-          const label = opt.label || opt.id
-          if (Object.keys(flattenInterventions(opt.interventions)).length === 0) {
-            emptyOptions.push({ id: opt.id, label })
-            continue
-          }
-          const unusable = unusableByOptionId.get(opt.id)
-          if (unusable && unusable.length > 0) {
-            partialOptions.push({ id: opt.id, label, targets: unusable })
-          }
-        }
-
-        const affectedOptions: Array<{ id: string; label: string }> = [
-          ...emptyOptions,
-          ...partialOptions.map((o) => ({ id: o.id, label: o.label })),
-        ]
-
-        if (affectedOptions.length > 0) {
-          const labels = emptyOptions.map((o) => o.label)
-          const emptyMessage =
-            emptyOptions.length === 0
-              ? ''
-              : emptyOptions.length === 1
-                ? `Option "${labels[0]}" needs intervention values before analysis can run.`
-                : `Options ${labels.map((o) => `"${o}"`).join(', ')} need intervention values before analysis can run.`
-          const partialMessage = partialOptions
-            .map((o) => {
-              const named = o.targets.map((t) => `"${targetName(t)}"`).join(', ')
-              const noun = o.targets.length === 1 ? 'an effect' : 'effects'
-              return `Option "${o.label}" has ${noun} on ${named} with no usable value. Set a number for each, or remove them, before analysis can run.`
-            })
-            .join(' ')
-          const message = [emptyMessage, partialMessage].filter(Boolean).join(' ')
-
-          if (import.meta.env.DEV) {
-            console.warn('[useV2Run] Pre-run validation failed: missing interventions', {
-              missingOptions: labels,
-              partiallyUnusableOptions: partialOptions.map((o) => ({
-                option: o.label,
-                targets: o.targets,
-              })),
-              usingAnalysisReady: !!effectiveAnalysisReady,
-            })
-          }
-
-          resultsError({
-            code: 'MISSING_INTERVENTIONS',
-            message,
-            request_id: requestId,
-            canRetry: false,
-            affectedOptions,
-          })
-
-          if (persistence) {
-            persistence.persistAnalysisFailure({
-              code: 'MISSING_INTERVENTIONS',
-              message,
-            }).catch(() => { /* non-critical */ })
-          }
-
-          setError(message)
-          setIsRunning(false)
-          return
-        }
-      }
-
-      if (import.meta.env.DEV) {
-        console.warn('[useV2Run] Starting V2 analysis', {
-          nodeCount: nodes.length,
-          edgeCount: edges.length,
-          goalNodeId: outcomeNodeId,
-          goalThreshold,
-          usingAnalysisReady: !!effectiveAnalysisReady,
-        })
-      }
-
-      // Execute V2 run with analysisReady (or fallback to node extraction)
-      // P0 Fix: Pass computed seed to avoid hardcoded "42" default
-      // Audit F-01: framing removed from PLoT request (PLoT rejects it with 400).
-      // P0 Fix: Include brief for PLoT context
-      // Lane 5: the V2 leg resolves the threshold through the SAME function as
-      // the V5 chip legs (resolveChipGoalThreshold) so the two can never
-      // disagree — same goal-node resolution (validated to exist), same cap
-      // chain (analysis_ready → node → saved-measure unit, provenance-guarded),
-      // and the same representation short-circuit (a bare-synced 'normalised'
-      // value passes through, never divided by the node cap → the 0.6/100 =
-      // 0.006 live corruption). UI-SEM-058/081 semantics preserved.
-      // Lane 5 (Codex P1): resolve the goal node ONCE and use it for BOTH the
-      // cap lookup and the request's outcome/goal node, so the threshold can
-      // never be normalised against one node's cap while the request names
-      // another. Prefers analysis_ready.goal_node_id (what the adapter also
-      // prefers), validates existence, falls back to outcome/first-goal.
-      const activeGoalNodeId = resolveActiveGoalNodeId(useCanvasStore.getState())
-      const measureForScenario = selectSuccessMeasure(
-        useSuccessMeasureStore.getState(),
-        resolveScenarioKey(useCanvasStore.getState().currentScenarioId),
-      )
-      const normalisedGoalThreshold = resolveChipGoalThreshold(goalThreshold, {
-        analysisReady: effectiveAnalysisReady as { goal_threshold_cap?: unknown } | null,
-        nodes,
-        goalNodeId: activeGoalNodeId,
-        unitCap: resolveMeasureUnitCap(measureForScenario, goalThreshold),
-        representation: goalThresholdRepresentation,
-      })
-      // A store threshold that EXISTS but cannot be proven normalised must
-      // CLEAR the request threshold (null → adapter deletes the builder's
-      // baked analysisReady.goal_threshold): the baked value predates the
-      // user's edit, and computing probability_of_goal against a replaced
-      // target silently misleads. No store value at all (undefined) leaves
-      // the builder's value standing.
-      const requestGoalThreshold =
-        normalisedGoalThreshold !== undefined
-          ? normalisedGoalThreshold
-          : goalThreshold != null && Number.isFinite(goalThreshold)
-            ? null
-            : undefined
-
-      const result = await executeV2RunWithAnalysisReady(
-        config,
-        nodes,
-        edges,
-        effectiveAnalysisReady,
-        // Lane 5 (Codex P1): the SAME resolved goal node the cap used above.
-        activeGoalNodeId ?? outcomeNodeId,
-        requestId,
-        requestGoalThreshold,
-        seed,
-        lastDraftDescription || undefined,
-        effectiveGoalConstraints,
-        framing?.scenario_id ?? null
-      )
-
-      const elapsed_ms = Date.now() - startTime
-
-      // Handle blocked response (422)
-      if (isBlockedResponse(result)) {
-        const errorResult = result as V2RunError
-        if (import.meta.env.DEV) {
-          console.warn('[useV2Run] Analysis blocked', {
-            requestId,
-            serverRequestId: errorResult.request_id,
-            reason: errorResult.status_reason,
-          })
-        }
-
-        // Surface PLoT EMPTY_INTERVENTIONS / MISSING_INTERVENTIONS critiques as
-        // a coached recovery state rather than a generic blocker. The coached
-        // UI is always preferable to "Something went wrong" for these codes,
-        // so we render it even when affected_nodes references a stale id that
-        // is no longer on the canvas — falling back to the id as the visible
-        // label gives the user something concrete to act on, and is more
-        // honest than the generic banner. Affected option ids come from
-        // critique.affected_nodes (PLoT references option node id directly).
-        const interventionCritiques = (errorResult.critiques || []).filter(
-          (c) => c.code === 'EMPTY_INTERVENTIONS' || c.code === 'MISSING_INTERVENTIONS',
-        )
-
-        let affectedOptions: Array<{ id: string; label: string }> | undefined
-        if (interventionCritiques.length > 0) {
-          const optionNodeById = new Map<string, { id: string; label: string }>()
-          for (const n of nodes) {
-            const d = (n.data ?? {}) as Record<string, unknown>
-            if (d.kind === 'option' || d.type === 'option') {
-              optionNodeById.set(n.id, {
-                id: n.id,
-                label: (d.label as string | undefined) || n.id,
-              })
-            }
-          }
-          const affectedIds: string[] = []
-          const seenIds = new Set<string>()
-          for (const crit of interventionCritiques) {
-            for (const id of crit.affected_nodes ?? []) {
-              if (!seenIds.has(id)) {
-                seenIds.add(id)
-                affectedIds.push(id)
-              }
-            }
-          }
-          // Resolve to canvas labels where possible, fall back to id-as-label
-          // for stale references so the coached UI can still render.
-          const resolved = affectedIds.map(
-            (id) => optionNodeById.get(id) ?? { id, label: id },
-          )
-          if (resolved.length > 0) affectedOptions = resolved
-        }
-
-        // Promote the code whenever the coached banner can render at least
-        // one option (resolved or id-fallback). If affected_nodes was empty
-        // entirely, we have nothing to render and fall back to the generic
-        // VALIDATION_BLOCKED path — which is what happens for cycles, missing
-        // goal, and other unrelated 422s.
-        // ROADMAP 1.54 (density wall): GRAPH_TOO_COMPLEX gets its own code so
-        // userFriendlyErrors renders the honest simplify-your-model copy —
-        // the producer's message names engine budget maths and must not be
-        // echoed (its raw text stays in the debug panel).
-        const complexityBlocked = (errorResult.critiques || []).some(
-          (c) => c.code === 'GRAPH_TOO_COMPLEX',
-        )
-        const promotedCode =
-          affectedOptions && affectedOptions.length > 0
-            ? interventionCritiques[0].code
-            : complexityBlocked
-              ? 'GRAPH_TOO_COMPLEX'
-              : 'VALIDATION_BLOCKED'
-
-
-        resultsError({
-          code: promotedCode,
-          message: errorResult.status_reason,
-          request_id: requestId,
-          canRetry: false, // User needs to fix model first
-          affectedOptions,
-        })
-
-        // C.1b: Persist failure to Supabase (non-blocking)
-        if (persistence) {
-          persistence.persistAnalysisFailure({
-            code: promotedCode,
-            message: errorResult.status_reason,
-          }).catch(() => { /* non-critical */ })
-        }
-
-        setError(errorResult.status_reason)
-        setIsRunning(false)
-        return
-      }
-
-      // Handle failed analysis (200 but failed)
-      if (isFailedAnalysis(result)) {
-        const failedResult = result as V2RunResponse
-        if (import.meta.env.DEV) {
-          console.warn('[useV2Run] Analysis failed', {
-            requestId,
-            serverRequestId: failedResult.request_id,
-          })
-        }
-
-
-        // Create error report with critiques
-        const errorReport = createErrorReport(
-          'Analysis could not complete. The model may have numerical instability.',
-          failedResult.critiques ?? [],
-          { seed }
-        )
-
-        resultsComplete({
-          report: errorReport,
-          hash: failedResult.response_hash,
-        })
-
-        // C.1b: Persist failure to Supabase (non-blocking)
-        if (persistence) {
-          persistence.persistAnalysisFailure({
-            code: 'ANALYSIS_FAILED',
-            message: 'Analysis could not complete',
-          }).catch(() => { /* non-critical */ })
-        }
-
-        setError('Analysis could not complete')
-        setIsRunning(false)
-        return
-      }
-
-      // Success - map V2 response to ReportV1
-      if (isSuccessfulAnalysis(result)) {
-        const rawSuccessResult = result as V2RunResponse
-
-        // Validate and sanitize before mapping to catch malformed responses
-        const validationResult = validateV2RunResponseFull(rawSuccessResult)
-        if (validationResult.softWarnings.length > 0 && import.meta.env.DEV) {
-          console.warn('[useV2Run] Soft validation warnings (will sanitize):', {
-            requestId,
-            warnings: validationResult.softWarnings,
-          })
-        }
-
-        // Sanitize to fix soft anomalies (e.g., object instead of array)
-        const successResult = sanitizeV2RunResponse(rawSuccessResult)
-
-        if (import.meta.env.DEV) {
-          console.warn('[useV2Run] Analysis complete', {
-            requestId,
-            serverRequestId: successResult.request_id,
-            status: successResult.analysis_status,
-            responseHash: successResult.response_hash,
-          })
-        }
-
-        const report = mapV2ResponseToReportV1(successResult, {
-          seed,
-          elapsed_ms,
-        })
-
-        // Create enrichment from V2 response for gate-state and useRobustness
-        const enrichment = createEnrichmentFromV2Response(successResult)
-
-        // Synthesize CEE V1 data from V2 response for Decision Review panel
-        // This provides useful AI-generated review content when using V2 endpoint
-        const ceeReviewV1 = synthesizeCeeReviewFromV2(successResult)
-        const ceeTraceV1 = synthesizeCeeTraceFromV2(successResult, requestId, elapsed_ms)
-
-        // Debug: Log raw M1 fields from V2 response before extraction
-        if (import.meta.env.DEV) {
-          console.warn('[useV2Run] Raw M1 Review fields from V2 response:', {
-            cee_status: successResult.cee_status,
-            has_decision_quality: !!successResult.decision_quality,
-            insights_count: successResult.insights?.length ?? 'missing',
-            guidance_count: successResult.improvement_guidance?.length ?? 'missing',
-            has_rationale: !!successResult.rationale,
-            rationale_summary: successResult.rationale?.summary?.slice(0, 80),
-            has_robustness: !!successResult.robustness_synthesis,
-            cee_trace: successResult.cee_trace ? {
-              request_id: successResult.cee_trace.request_id,
-              latency_ms: successResult.cee_trace.latency_ms,
-              degraded: successResult.cee_trace.degraded,
-            } : 'missing',
-          })
-        }
-
-        // Extract M1 Review data (rationale, robustness synthesis, etc.)
-        const m1Review = extractM1ReviewFromV2(successResult)
-
-        // Extract M1 Coaching data (deterministic coaching fields)
-        const m1Coaching = extractM1CoachingFromV2(successResult)
-
-        // Extract M1 Review assumptions (key_assumptions + pre_mortem from PLoT)
-        const m1ReviewAssumptions = successResult.m1_review ?? null
-
-        if (import.meta.env.DEV) {
-          console.warn('[useV2Run] Extracted M1 Review:', {
-            hasCeeReview: !!ceeReviewV1,
-            blockCount: ceeReviewV1?.blocks?.length ?? 0,
-            readinessLevel: ceeReviewV1?.readiness?.level,
-            hasM1Review: !!m1Review,
-            m1CeeStatus: m1Review?.cee_status,
-            hasRationale: !!m1Review?.rationale,
-            hasRobustnessSynthesis: !!m1Review?.robustness_synthesis,
-            insightsCount: m1Review?.insights?.length ?? 0,
-            guidanceCount: m1Review?.improvement_guidance?.length ?? 0,
-          })
-          console.warn('[useV2Run] Extracted M1 Coaching:', {
-            hasM1Coaching: !!m1Coaching,
-            hasHeadline: !!m1Coaching?.executive_summary?.headline,
-            readiness: m1Coaching?.readiness,
-            storyHeadlinesCount: Object.keys(m1Coaching?.story_headlines ?? {}).length,
-            evidenceGapsCount: m1Coaching?.evidence_gaps?.length ?? 0,
-            nextActionsCount: m1Coaching?.next_actions?.length ?? 0,
-            dominantFactor: m1Coaching?.key_drivers?.dominant_factor,
-          })
-        }
-
-        // Detect computed-but-empty anomalies (backend bug detection)
-        const anomalies = detectComputedButEmpty(successResult)
-        if (anomalies.length > 0) {
-          // Track the anomaly for debugging
-          trackEmptyComputedResults({
-            request_id: requestId,
-            anomalies: anomalies.map(a => ({
-              field: a.field,
-              status: a.status,
-              message: a.message,
-            })),
-          })
-
-          if (import.meta.env.DEV) {
-            console.warn('[useV2Run] Detected computed-but-empty anomalies', {
-              requestId,
-              anomalies,
-            })
-          }
-        }
-
-        // Pass synthesized CEE V1 data to store so Decision Review panel can display content
-        resultsComplete({
-          report,
-          hash: successResult.response_hash,
-          requestId,
-          enrichment,
-          ceeReviewV1,
-          ceeTraceV1,
-          rawV2Response: rawSuccessResult,
-        })
-
-        try {
-          useResultsStore.getState().setAnalysisSummary(assembleAnalysisInputsSummary(successResult))
-        } catch {
-          useResultsStore.getState().setAnalysisSummary(null)
-        }
-
-        // C.1b: Persist analysis results to Supabase (non-blocking)
-        if (persistence) {
-          // T2b: the seed has TWO sinks here, and they answer DIFFERENT questions.
-          //
-          // 1. `seedUsed` is the RECEIPT: "what seed did the engine say it used?"
-          //    Unknown must stay unknown (null) — see resolveSeedUsed. This is
-          //    what lands in Supabase provenance and is read back by
-          //    hydrateAnalysis on reload, so a fabricated 0 here becomes a
-          //    real-looking "Seed 0" row (the exact bug PR #326 removed from
-          //    the read path).
-          //
-          // 2. `seedForHash` is a RUN IDENTITY for the graph hash, not a claim
-          //    about the engine. When the engine did not echo a usable seed we
-          //    fall back to the seed we actually SENT (`seed` is always a real
-          //    number by this point — derived from framing.seed, else hashed
-          //    from scenario_id, else a timestamp fallback; see deriveSeed
-          //    just below the goal validation), because
-          //    that is the seed a replay would resend. Falling back to the
-          //    requested seed here is strictly more honest than the old code,
-          //    which hashed a fabricated 0 whenever the echo was malformed.
-          //
-          // The hash's seed component cannot itself distinguish a real 0 from
-          // "unknown" — computeClientHash (adapters/plot/v1/mapper.ts:256) does
-          // `seed: seed || 0`, collapsing null/undefined/NaN/0 to the same
-          // canonical value. Fixing that means changing hash semantics for the
-          // replay gate (pinned by adapters/__tests__/wave2-replay-gate.spec.ts)
-          // and for two out-of-lane call sites (store.ts:2904,
-          // ReactFlowGraph.tsx:1532/1611), so it is deliberately left alone and
-          // ledgered rather than changed casually here.
-          const seedUsed = resolveSeedUsed(successResult.meta?.seed_used)
-          const seedForHash = seedUsed ?? seed
-          const graphHash = generateGraphHash(nodes, edges, seedForHash)
-          persistence.persistAnalysisSuccess(
-            successResult,
-            graphHash,
-            seedUsed,
-            successResult.response_hash,
-            {
-              option_count: successResult.option_comparison?.length ?? 0,
-              analysis_status: successResult.analysis_status,
-            },
-          ).catch((err) => {
-            if (import.meta.env.DEV) {
-              console.warn('[useV2Run] Supabase analysis persistence failed', err)
-            }
-          })
-        }
-
-        // Update run metadata with CEE data for debugging/display
-        setRunMeta({
-          ceeReviewV1,
-          ceeTraceV1,
-          ceeErrorV1: null, // No error - synthesis succeeded
-          m1Review, // M1 Review enrichment (rationale, robustness synthesis, etc.)
-          m1Coaching, // M1 Coaching (deterministic, not LLM-generated)
-          m1ReviewAssumptions, // M1 Review assumptions + pre-mortem from PLoT
-          reviewStatus: successResult.review_status, // V12: M2 gate
-        })
-
-        // Update gates after successful analysis
-        // BUT: If there are anomalies, set gate to 'warn' instead of 'pass'
-        const hasRobustnessAnomaly = anomalies.some(a => a.field === 'robustness')
-        const hasOptionAnomaly = anomalies.some(a => a.field === 'option_comparison')
-
-        if (hasOptionAnomaly) {
-          // Option comparison claimed "computed" but was empty - this is a serious issue
-          useGateStore.getState().setGate('run', 'warn', {
-            requestId,
-            message: 'Analysis completed but option results are missing',
-            service: 'plot-v2',
-          })
-        } else {
-          useGateStore.getState().setGate('run', 'pass', {
-            requestId,
-            message: 'Analysis completed successfully',
-            service: 'plot-v2',
-          })
-        }
-
-        // Update robustness gate from V2 response directly
-        // The new function handles robustness_status and robustness data properly
-        // Special case: hasRobustnessAnomaly means computed but empty (backend bug)
-        if (hasRobustnessAnomaly) {
-          useGateStore.getState().setGate('robustness', 'warn', {
-            requestId,
-            message: 'Robustness computed but results are empty',
-            service: 'plot-v2',
-          })
-        } else {
-          // Use the new V2-aware gate function that accepts robustness_status directly
-          updateRobustnessGateFromV2(
-            successResult.robustness_status,
-            successResult.robustness
-          )
-        }
-
-        // Show partial warning if applicable
-        if (result.analysis_status === 'partial') {
-          console.warn('[useV2Run] Partial results returned')
-        }
-
-        setIsRunning(false)
-        return
-      }
-
-      // Unexpected state
-      console.error('[useV2Run] Unexpected result state', result)
-      resultsError({
-        code: 'UNEXPECTED_STATE',
-        message: 'Received unexpected response format',
-        request_id: requestId,
-        canRetry: true, // Might be transient
-      })
-      setError('Unexpected response format')
-    } catch (err) {
-      // P0 race fix: If this run was superseded (a newer run owns the controller),
-      // no-op — the newer run manages state. Only the active run may mutate status.
-      const isActiveRun = abortControllerRef.current === controller
-
-      // Distinguish user-initiated cancellation from other abort errors
-      if (controller.signal.aborted && err instanceof Error && err.name === 'AbortError') {
-        // Only fire resultsCancelled for user-initiated cancel on the active run.
-        // Superseded runs (aborted by F-77 guard) must not touch state.
-        if (isActiveRun) {
-          resultsCancelled()
-          setIsRunning(false)
-          abortControllerRef.current = null
-          // Reset Supabase analysis_status back to 'none' (fire-and-forget)
-          persistence?.resetAnalysisStatus?.().catch(() => {})
-        }
-        return
-      }
-
-      // P0 race fix: If this run was superseded, do not mutate shared state.
-      // `trackTypedError` below is still safe — it's fire-and-forget. (The
-      // `trackRunFailed` this comment used to name moved to OutputsDock; see
-      // the run-spine note at the top of this file.)
-      if (!isActiveRun) {
-        return
-      }
-
-      // `elapsed_ms` was computed here only to be passed to `trackRunFailed`,
-      // which no longer emits from this hook. Removed rather than left unused:
-      // a dead duration is how a future reader concludes the timing is still
-      // being measured here when it is measured in OutputsDock.
-      const message = err instanceof Error ? err.message : 'Unknown error'
-
-      if (import.meta.env.DEV) {
-        console.error('[useV2Run] Error', { requestId, error: err })
-      }
-
-      // Convert to typed error for consistent handling
-      let typedError: ApiError
-      if (isApiError(err)) {
-        // Already a typed error (e.g., MalformedApiResponseError from adapter)
-        typedError = err
-      } else if (err instanceof Error && (
-        err.name === 'AbortError' ||
-        err.name === 'TypeError' && message.includes('fetch') ||
-        message.includes('NetworkError') ||
-        message.includes('Failed to fetch') ||
-        message.includes('network') ||
-        message.includes('timeout') ||
-        message.includes('Connection refused')
-      )) {
-        // Network-related error
-        typedError = NetworkError.fromFetchError(err, { requestId, endpoint: '/v2/run' })
-      } else {
-        // Processing/mapping error
-        typedError = ProcessingError.wrap(err, { requestId, endpoint: '/v2/run' })
-      }
-
-      const errorCode = typedError.code
-
-      // Track typed error for observability
-      trackTypedError(typedError)
-
-
-      resultsError({
-        code: errorCode,
-        message,
-        request_id: requestId,
-        canRetry: typedError.retryable,
-      })
-
-      // C.1b: Persist failure to Supabase (non-blocking)
-      if (persistence) {
-        persistence.persistAnalysisFailure({
-          code: errorCode,
-          message,
-        }).catch(() => { /* non-critical */ })
-      }
-
-      // Capture error detail for debug drawer expansion
-      captureErrorDetail({
-        timestamp: new Date().toISOString(),
-        service: 'PLoT',
-        httpStatus: (typedError as NetworkError).status,
-        errorCode: typedError.code,
-        message,
-        requestId,
-        endpoint: '/v2/run',
-        retryable: typedError.retryable,
-        rawBody: typedError.context.rawPayload
-          ? JSON.stringify(typedError.context.rawPayload).slice(0, 2000)
-          : undefined,
-      })
-
-      // Capture raw error data for debugging in the debug panel
-      // Uses redaction utility to protect privacy and enforce size limits
-      let rawErrorPayload: unknown = undefined
-      let validationErrors: string[] | undefined
-
-      // If this is a MalformedApiResponseError, capture validation details
-      if (err && typeof err === 'object' && 'validationErrors' in err) {
-        const malformedErr = err as {
-          validationErrors?: Array<{ field: string; expected: string; received: string }>
-          rawPayload?: unknown
-        }
-        rawErrorPayload = malformedErr.rawPayload
-        validationErrors = malformedErr.validationErrors?.map(
-          e => `${e.field}: expected ${e.expected}, got ${e.received}`
-        )
-      }
-
-      // Build redacted error data (respects dev/staging gating and size limits)
-      const rawErrorData = buildRawErrorData({
-        payload: rawErrorPayload,
-        expectedShape: 'V2RunResponse | V2RunError',
-        validationErrors,
-      })
-
-      // Store raw error data for debug panel access
-      setRunMeta({ rawErrorData })
-
-      setError(message)
-    } finally {
-      // P0 race fix: Only the active run clears shared state.
-      // A superseded run must not reset isRunning or the controller ref,
-      // as a newer run now owns them.
-      if (abortControllerRef.current === controller) {
-        abortControllerRef.current = null
-        setIsRunning(false)
-      }
-    }
-  // Audit F-76: Dependency array reduced — graph state is read from getState() inside callback.
-  // Only persistence (external prop) can change the callback's behaviour.
-  }, [persistence])
-
-  const cancelRun = useCallback(() => {
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort()
-    }
-  }, [])
-
-  return {
-    runV2Analysis,
-    cancelRun,
-    isRunning,
-    error,
-  }
-}
